backend/requirements.txt
# Existing Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
werkzeug==2.3.7
requests==2.31.0

# File processing
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
uuid==1.30

# AI/ML dependencies (from training.ipynb)
torch>=1.9.0
librosa>=0.8.0
soundfile>=0.10.0
numpy>=1.20.0
pretty_midi>=0.2.9

# Optional: for better performance
numba>=0.56.0  # Used by librosa for speed

# Development/testing (optional)
matplotlib>=3.5.0  # For visualization 
tqdm>=4.60.0      # Progress bars